+++ demo/App.svelte
@@ -25,9 +25,11 @@ init(routes);
 
 <!-- Anchor navigation & using route.path -->
 <div>
-	<a href="/welcome/john" {@attach link} class:active={route.path === '/welcome/john'}>Welcome John</a>
+	<!-- Without params -->
+	<a href="/welcome/john" {@attach link()} class:active={route.path === '/welcome/john'}>Welcome John</a>
 	|
-	<a href="/welcome/jane" {@attach link} class:active={route.path === '/welcome/jane'}>Welcome Jane</a>
+	<!-- With params -->
+	<a href="/welcome/:name" {@attach link({ params: { name: 'jane' } })} class:active={route.path === '/welcome/jane'}>Welcome Jane</a>
 </div>
 
 <!-- Router outlet -->