+++ src/index.d.ts
@@ -33,6 +33,22 @@ export const init: (
 
 ) => Promise<void>;
 
+export type NavigateOptions = {
+
+	/**
+	 * If `true`, `History.replaceState()` will be used, otherwise `History.pushState()`
+	 * @default false
+	 */
+	replace?: boolean,
+
+	/** Params to inject into url */
+	params?: { [key: string]: string },
+
+	/** Query params to add after path */
+	query?: { [key: string]: string }
+
+};
+
 /** Navigate to a path */
 export const navigate: (
 
@@ -40,21 +56,7 @@ export const navigate: (
 	path: string,
 
 	/** Options */
-	options?: {
-
-		/**
-		 * If `true`, `History.replaceState()` will be used, otherwise `History.pushState()`
-		 * @default false
-		 */
-		replace?: boolean,
-
-		/** Params to inject into url */
-		params?: { [key: string]: string },
-
-		/** Query params to add after path */
-		query?: { [key: string]: string }
-
-	}
+	options?: NavigateOptions
 
 ) => Promise<void>;
 
@@ -75,5 +77,10 @@ export const route: {
 
 };
 
-/** Attachment that can be used on `a` elements to change them from regular links to router links */
-export const link: Attachment<HTMLAnchorElement>;
+/** Attachment factory that can be used on `a` elements to change them from regular links to router links */
+export const link: (
+
+	/** Options */
+	options?: NavigateOptions
+
+) => Attachment<HTMLAnchorElement>;